requirements_dev.txt
pyright==1.1.402
black==25.1.0
isort==6.0.1
asyncpg-stubs==0.30.2
